requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
PyJWT
bcrypt
pymysql
alembic
pytest
pydantic
flask_restx
faker
email-validator
gunicorn